Fix timer scaling. Thanks to John Wright.

+++ RNTimer/RNTimer.m
@@ -45,7 +45,7 @@
   timer.block = block;
   timer.source = dispatch_source_create(DISPATCH_SOURCE_TYPE_TIMER, 0, 0,
       dispatch_get_main_queue());
-  uint64_t nsec = (unsigned)(seconds * NSEC_PER_SEC);
+  uint64_t nsec = (uint64_t)(seconds * NSEC_PER_SEC);
   dispatch_source_set_timer(timer.source, dispatch_time(DISPATCH_TIME_NOW, nsec),
                             nsec, 0);
   dispatch_source_set_event_handler(timer.source, block);
